+++ CHANGES.rst
@@ -40,6 +40,7 @@ Bugfixes
 - Fix event header bar in IE11 (:issue:`3135`)
 - Fix footer on login page (:issue:`3132`)
 - Use correct module name for abstract notification emails in the event log
+- Remove linebreaks from email subject in paper review notifications
 
 
 Version 2.0a1